Add README.md, index.html, and tortilla-soup.html Filled README.md with project requirements, a baseline plan of how I'll go about completing the project, and what I will take away from the project.
Created index.html, the homepage of my project, and placed an unordered list of recipes within it.

Created tortilla-soup.html, added a placeholder for the image of the dish, a short description of the dish, and ingredients. All ingredients fall within unordered lists, whereas directions will fall within an ordered list. Ingredients are separated by heading elements, as different ingredients are used for different parts of the dish.

To-do:

Create the two additional pages for recipes
Add the Chicken Tortilla Soup image
Add credits to the original recipe on each webpage
Add links to each recipe on the main index.html page
Add a link back to the main index.html page on each recipe page, including tortilla-soup.html