Finished directions for Chicken Tortilla Soup
Added an ordered list under directions header, detailing the instructions required to make the Chicken Tortilla Soup dish.

Added <strong> elements throughout directions for readability; in the parts where attention shifts to preparing another part of the dish, e.g. going from the broth to preparing the tortilla strips, a strong element is added. Helps screen readers read the instructions properly as well.

If not a strong element, the section is marked in such a manner: "For Salsa:", or "For Tortilla Strips:"