Created template for recipes
Created a resources folder to house future resources if necessary.

Inside of that resources folder, there is a file named "placeholder.html". It serves as a template to streamline the process of adding further recipes to the website.

Rather than having to write out all the code again, I can just insert the information into my elements and be at ease.